refactor: extract CountriesList
