+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter, Redirect, Route, Switch } from 'react-router-dom';
 import Footer from './components/Footer/Footer';
 import CountryPage from './pages/CountryPage';
-import MainPage from './pages/MainPage/';
+import MainPage from './pages/MainPage';
 
 const App = () => {
   return (
