+++ src/pages/MainPage.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import Header from '../components/Header/Header';
+import Search from '../components/Search/Search';
+import Wrapper from '../components/Wrapper/Wrapper';
+import CountriesList from '../components/CountriesList/CountriesList';
+
+const MainPage = () => (
+  <React.Fragment>
+    <Header>
+      <Search />
+    </Header>
+    <Wrapper>
+      <CountriesList />
+    </Wrapper>
+  </React.Fragment>
+);
+
+export default MainPage;
